Azure/ARM Templates/Sentinel and Google Alert Center Integration/requirements.txt
azure-functions
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
azure-storage-file-share==12.3.0